合并装备和发布的脚本。

+++ make_image.sh
@@ -17,7 +17,7 @@ function check_ip() {
     fi
 }
 
-if [ $# != 5 ] ; then
+if [ $# != 6 ] ; then
     echo "Make image packe script param err!"
     exit 1
 fi
@@ -49,10 +49,27 @@ cd GKIPCLinuxV100R001C00SPC030
 source build/env.sh
 cd ..
 
+make_cmd=""
+if [[ ${6} = "eqm" ]]
+then
+    make_cmd="make eqm"
+fi
+
+if [[ ${6} = "release" ]]
+then
+    make_cmd="make"
+fi
+
+if [[ ${make_cmd} = "" ]]
+then
+    echo "Please input correct type, eqm or release!"
+    exit -1
+fi
+
 cd sample
 make clean
-make
-
+echo $make_cmd
+$make_cmd
 cd ../
 
 rm -rf ./rootfs_package
@@ -83,4 +100,11 @@ arm-gcc6.3-linux-uclibceabi-strip $packageAlgDir/search
 current=`date "+%Y%m%d%H%M%S"`
 
 chmod +x ./mkfs.jffs2
-./mkfs.jffs2 -d ./rootfs_package -l -e 0x10000 -o rootfs-$current.jffs2
+mkfs_cmd="./mkfs.jffs2 -d ./rootfs_package -l -e 0x10000 -o rootfs-$current.jffs2"
+
+if [[ ${6} = "eqm" ]]
+then
+    mkfs_cmd="./mkfs.jffs2 -d ./rootfs_package -l -e 0x10000 -o rootfs-eqm-$current.jffs2"
+fi
+echo $mkfs_cmd
+$mkfs_cmd
